+++ CHANGELOG.md
@@ -13,10 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-### Fixes
+### Fixed
 
 - Fix authentication-method not being set after upgrading to 2.9.0 [#833](https://github.com/nextcloud/integration_openproject/pull/833)
 - Fix OpenProject icon for proper rendering [#835](https://github.com/nextcloud/integration_openproject/pull/834)
+- Persist authentication settings form state after save [#827](https://github.com/nextcloud/integration_openproject/pull/827)
 
 ## 2.9.0 - 2025-05-21